Remove unused code, move WarningDialog to top level of components
