--- src/components/ExternalLoading/ExternalNetworkLoadingPanel.tsx
@@ -1,86 +1,0 @@
-import { ReactElement, useContext, useEffect } from 'react'
-import { Box, LinearProgress } from '@mui/material'
-import { useLocation, useNavigate } from 'react-router-dom'
-import { Workspace } from '../../models/WorkspaceModel'
-import { useWorkspaceStore } from '../../store/WorkspaceStore'
-import { useCredentialStore } from '../../store/CredentialStore'
-import { getSummariesFromCacheOrNdex } from '../../store/getNetworkSummaryFromCacheOrNdex'
-import { AppConfigContext } from '../../AppConfigContext'
-import { useUiStateStore } from '../../store/UiStateStore'
-import { logUi } from '../../debug'
-
-interface ExternalNetworkLoadingPanelProps {
-  message: string
-  showProgress?: boolean
-}
-
-export const ExternalNetworkLoadingPanel = (
-  props: ExternalNetworkLoadingPanelProps,
-): ReactElement => {
-  const { ndexBaseUrl } = useContext(AppConfigContext)
-  const navigate = useNavigate()
-  const workspace: Workspace = useWorkspaceStore((state) => state.workspace)
-  const { id } = workspace
-  const setShowErrorDialog = useUiStateStore(
-    (state) => state.setShowErrorDialog,
-  )
-
-  const setErrorMessage = useUiStateStore((state) => state.setErrorMessage)
-
-  const getToken: () => Promise<string> = useCredentialStore(
-    (state) => state.getToken,
-  )
-
-  const addNetworkIds: (networkId: string) => void = useWorkspaceStore(
-    (state) => state.addNetworkIds,
-  )
-
-  const setCurrentNetworkId: (networkId: string) => void = useWorkspaceStore(
-    (state) => state.setCurrentNetworkId,
-  )
-
-  const location = useLocation()
-  const redirect = (): void => {
-    const networkId = location.pathname.split('/')[2]
-    void getToken().then((token) => {
-      getSummariesFromCacheOrNdex(networkId, ndexBaseUrl, token)
-        .then((summary) => {
-          // Add to the workspace
-          addNetworkIds(networkId)
-          setCurrentNetworkId(networkId)
-          navigate(`/${id}/networks/${networkId}${location.search.toString()}`)
-        })
-        .catch((error) => {
-          logUi.error(
-            `[${ExternalNetworkLoadingPanel.name}]:[${redirect.name}]: Failed to load the network ${networkId}`,
-            error,
-          )
-          const errorMessage: string = error.message
-          setErrorMessage(
-            `Failed to load the network ${networkId}: ${errorMessage}`,
-          )
-          setShowErrorDialog(true)
-        })
-    })
-  }
-
-  useEffect(() => {
-    logUi.info(
-      `[${ExternalNetworkLoadingPanel.name}]:[${useEffect.name}]: workspace: ${workspace}`,
-    )
-    if (id !== undefined) {
-      redirect()
-    }
-  }, [id])
-
-  return (
-    <Box
-      sx={{ width: '100%', height: '100%', display: 'grid', padding: '1em' }}
-    >
-      <Box sx={{ margin: 'auto' }}>
-        <h2>{props.message}</h2>
-        {(props.showProgress ?? false) ? <LinearProgress /> : null}
-      </Box>
-    </Box>
-  )
-}
