--- src/components/ExternalLoading/index.tsx
@@ -1,1 +1,0 @@
-export { ExternalNetworkLoadingPanel } from './ExternalNetworkLoadingPanel'
